+++ pkg/aws/v1alpha3/machinedeployment/validate_machinedeployment_test.go
@@ -69,3 +69,77 @@ func TestValidateCluster(t *testing.T) {
 		})
 	}
 }
+
+func TestValidateClusterNamespace(t *testing.T) {
+	testCases := []struct {
+		ctx  context.Context
+		name string
+
+		clusterNamespace  string
+		nodePoolNamespace string
+		allowed           bool
+	}{
+		{
+			ctx:  context.Background(),
+			name: "case 0",
+
+			clusterNamespace:  "default",
+			nodePoolNamespace: "default",
+			allowed:           true,
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 1",
+
+			clusterNamespace:  "org-test",
+			nodePoolNamespace: "org-test",
+			allowed:           true,
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 2",
+
+			clusterNamespace:  "default",
+			nodePoolNamespace: "org-test",
+			allowed:           false,
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 3",
+
+			clusterNamespace:  "org-test",
+			nodePoolNamespace: "default",
+			allowed:           false,
+		},
+	}
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			var err error
+
+			fakeK8sClient := unittest.FakeK8sClient()
+			validate := &Validator{
+				k8sClient: fakeK8sClient,
+				logger:    microloggertest.New(),
+			}
+
+			// create the cluster
+			cluster := unittest.DefaultCluster()
+			cluster.SetNamespace(tc.clusterNamespace)
+			err = fakeK8sClient.CtrlClient().Create(tc.ctx, cluster)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			// try to create the awsmachinedeployment
+			object := unittest.DefaultMachineDeployment()
+			object.SetNamespace(tc.nodePoolNamespace)
+			err = validate.ValidateCluster(object)
+			if tc.allowed && err != nil {
+				t.Fatalf("unexpected error %v", err)
+			}
+			if !tc.allowed && err == nil {
+				t.Fatalf("expected error but returned %v", err)
+			}
+		})
+	}
+}