+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added validation and tests for cluster CRs to be created in the organization-namespace from version `16.0.0`.
+- Added tests for nodepool CRs to be created in the same namespace as the cluster.
+
 ## [3.2.1] - 2021-08-12
 
 ### Added